iniciando a implementação de agendas

+++ database/migrations/2021_08_27_180629_create_table_projetos.php
@@ -18,7 +18,7 @@ class CreateTableProjetos extends Migration
             $table->timestamps();
             $table->string('nome');
             $table->string('descricao')->nullable();
-            $table->integer('id_projeto_pai')->nullable();
+            $table->unsignedBigInteger('id_projeto_pai')->nullable();
             $table->integer('nivel_projeto')->default(0);
             $table->dateTime('data_criacao')->default(date('Y-m-d h:i:s'));
             $table->bigInteger('custo_previsto')->default(0);
